Move some table util functions to _utils.py
